NewSidebar performance improvements (#274)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
N/A

**Description**
- Moved type to the proper file
- Moved style to the proper file
- Used useCallback hook to improve performance reducing unnecessary
renderings.
- Used useMemo hook to improve performance reducing unnecessary
processing.
- Unified NewSidebarItem and SidebarItem.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

**Checklist**

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [ ] This change is covered by unit tests
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated

---------
